remove faye handler
I realized it's a bad idea because pubsub doesn't distinguish
between different clients and sessions. Everytime that the faye
handler was opened, it would publish to every other connected client
rather than just the requesting client.

Down the line, I might explore adding a pubsub interface to rack-stream,
but it's out of scope for now.